:recycle: [Refactor] 불필요한 코드 삭제 및 파일 구조 변경
