+++ src/main/java/com/gamegoo/controller/member/JoinController.java
@@ -20,7 +20,7 @@ public class JoinController {
     @PostMapping("/join/local")
     @Operation(summary = "회원가입 API 입니다.", description = "API for join")
     public ApiResponse<Object> joinProcess(JoinDTO joinDTO) {
-        joinService.JoinProcess(joinDTO);
+        joinService.JoinMember(joinDTO);
         return ApiResponse.onSuccess(null);
     }
