+++ src/main/java/com/gamegoo/repository/MemberRepository.java
@@ -5,7 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Member findByEmail(String email);
-
-    Boolean existsByEmail(String email);
-
 }
