+++ src/main/java/com/gamegoo/service/member/CustomMemberDetailService.java
@@ -1,8 +1,8 @@
 package com.gamegoo.service.member;
 
 import com.gamegoo.domain.Member;
-import com.gamegoo.dto.member.CustomMemberDetails;
 import com.gamegoo.repository.MemberRepository;
+import com.gamegoo.security.CustomMemberDetails;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
